handle errors

+++ src/idea/idea.service.ts
@@ -1,9 +1,9 @@
-import { Injectable } from "@nestjs/common";
+import { Injectable, HttpException, HttpStatus } from "@nestjs/common";
 import { InjectRepository } from "@nestjs/typeorm";
 import { Repository } from "typeorm";
 
 import { IdeaEntity } from "./idea.entity";
-import { IdeaDIO } from "./dto/idea.dto";
+import { IdeaDTO } from "./dto/idea.dto";
 
 @Injectable()
 export class IdeaService {
@@ -16,7 +16,7 @@ export class IdeaService {
     return await this.ideaRepository.find();
   }
 
-  public async create(data: IdeaDIO) {
+  public async create(data: IdeaDTO) {
     const idea = await this.ideaRepository.create(data);
     await this.ideaRepository.save(idea);
 
@@ -24,18 +24,39 @@ export class IdeaService {
   }
 
   public async read(id: string) {
-    return await this.ideaRepository.findOne({ where: { id } });
+    const idea = await this.ideaRepository.findOne({ where: { id } });
+
+    if (!idea) {
+      throw new HttpException("Not found", HttpStatus.NOT_FOUND);
+    }
+
+    return idea;
   }
 
-  public async update(id: string, data: Partial<IdeaDIO>) {
+  public async update(id: string, data: Partial<IdeaDTO>) {
+    let idea = await this.ideaRepository.findOne({ where: { id } });
+
+    if (!idea) {
+      throw new HttpException("Not found", HttpStatus.NOT_FOUND);
+    }
+
     await this.ideaRepository.update({ id }, data);
 
-    return await this.ideaRepository.findOne({ where: { id } });
+    // return the value updated
+    idea = await this.ideaRepository.findOne({ where: { id } });
+
+    return idea;
   }
 
-  public async destroy(id: string): Promise<boolean> {
+  public async destroy(id: string) {
+    const idea = await this.ideaRepository.findOne({ where: { id } });
+
+    if (!idea) {
+      throw new HttpException("Not found", HttpStatus.NOT_FOUND);
+    }
+
     await this.ideaRepository.delete({ id });
 
-    return true;
+    return idea;
   }
 }
